move images to assets/images/
