+++ js/assets/images/contracts/gavcoin-64x64.png
