+++ js/assets/images/contracts/unknown-64x64.png
