+++ js/assets/images/ethcore-logo-white-square.png
