+++ js/assets/images/parity.xcf
