+++ js/src/dapps/signaturereg/Header/header.js
@@ -17,7 +17,7 @@
 import React, { Component, PropTypes } from 'react';
 
 import styles from './header.css';
-import blocks from '../images/blocks-350.png';
+import blocks from '../../../../assets/images/dapps/blocks-350.jpg';
 
 export default class Header extends Component {
   static propTypes = {
