+++ js/src/modals/Shapeshift/shapeshift.js
@@ -20,7 +20,7 @@ import ContentClear from 'material-ui/svg-icons/content/clear';
 
 import { Button, IdentityIcon, Modal } from '../../ui';
 import initShapeshift from '../../3rdparty/shapeshift';
-import shapeshiftLogo from '../../images/shapeshift-logo.png';
+import shapeshiftLogo from '../../../assets/images/shapeshift-logo.png';
 
 import AwaitingDepositStep from './AwaitingDepositStep';
 import AwaitingExchangeStep from './AwaitingExchangeStep';
